Confusion Matrix for Categorical Data

Calculates a cross-tabulation of observed and predicted classes.

For conf_mat() objects, the tidy method collapses the cell counts by cell into a data frame for each manipulation.

conf_mat(data, ...)

# S3 method for data.frame
conf_mat(data, truth, estimate, dnn = c("Prediction",
  "Truth"), ...)

# S3 method for table
conf_mat(data, ...)

# S3 method for conf_mat
tidy(x, ...)

Arguments

data
A data frame or a base::table().

...
Options to pass to base::table() (not including dnn). This argument is not currently used for the tidy method.

truth
The column identifier for the true class results (that is a factor). This should an unquoted column name although this argument is passed by expression and support quasiquotation (you can unquote column names or column positions).

estimate
The column identifier for the predicted class results (that is also factor). As with truth this can be specified different ways but the primary method is to use an unquoted variable name.

dnn
a character vector of dimnames for the table

x
A object of class conf_mat().

Value

conf_mat produces a object with class conf_mat. This contains the table and other objects. tidy.conf_mat generates a tibble with columns name (the cell identifier) and value (the cell count).

Details

The function requires that the factors have exactly the same levels.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

data("hpc_cv")

# The confusion matrix from a single assessment set (i.e. fold)
hpc_cv %>%
  filter(Resample == "Fold01") %>%
  conf_mat(obs, pred)

#>           Truth
#> Prediction  VF   F   M   L
#>         VF 166  33   8   1
#>         F   11  71  24   7
#>         M    0   3   5   3
#>         L    0   1   4  10

# Now compute the average confusion matrix across all folds in
# terms of the proportion of the data contained in each cell. 
# First get the raw cell counts per fold using the `tidy` method
cells_per_resample <- hpc_cv %>%
  group_by(Resample) %>%
  do(tidy(conf_mat(., obs, pred)))

# Get the totals per resample
counts_per_resample <- hpc_cv %>%
  group_by(Resample) %>%
  summarize(total = n()) %>%
  left_join(cells_per_resample, by = "Resample") %>%
  # Compute the proportions
  mutate(prop = value/total) %>%
  group_by(name) %>%
  # Average
  summarize(prop = mean(prop))

counts_per_resample

#> # A tibble: 16 x 2
#>        name         prop
#>       <chr>        <dbl>
#>  1 cell_1_1 0.4672451291
#>  2 cell_1_2 0.1070125709
#>  3 cell_1_3 0.0184655413
#>  4 cell_1_4 0.0025945025
#>  5 cell_2_1 0.0406866182
#>  6 cell_2_2 0.1866050267
#>  7 cell_2_3 0.0631650619
#>  8 cell_2_4 0.0173094431
#>  9 cell_3_1 0.0017341137
#> 10 cell_3_2 0.0069206198
#> 11 cell_3_3 0.0227791313
#> 12 cell_3_4 0.0080725199
#> 13 cell_4_1 0.0005747126
#> 14 cell_4_2 0.0103930022
#> 15 cell_4_3 0.0144251000
#> 16 cell_4_4 0.0320169067

# Now reshape these into a matrix
mean_cmat <- matrix(counts_per_resample$prop, byrow = TRUE, ncol = 4)
rownames(mean_cmat) <- levels(hpc_cv$obs)
colnames(mean_cmat) <- levels(hpc_cv$obs)

round(mean_cmat, 3)

#>       VF     F     M     L
#> VF 0.467 0.107 0.018 0.003
#> F  0.041 0.187 0.063 0.017
#> M  0.002 0.007 0.023 0.008
#> L  0.001 0.010 0.014 0.032