F Measure

These functions calculate the f_meas() of a measurement system for finding relevant documents compared to reference results (the truth regarding relevance). Highly related functions are recall() and precision().

f_meas(data, ...)

# S3 method for data.frame
f_meas(data, truth, estimate, beta = 1,
  estimator = NULL, na_rm = TRUE, ...)

f_meas_vec(truth, estimate, beta = 1, estimator = NULL, na_rm = TRUE,
  ...)

Arguments

data
Either a data.frame containing the truth and estimate columns, or a table/matrix where the true class results should be in the columns of the table.

...
Not currently used.

truth
The column identifier for the true class results (that is a factor). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a factor vector.

estimate
The column identifier for the predicted class results (that is also factor). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a factor vector.

beta
A numeric value used to weight precision and recall. A value of 1 is traditionally used and corresponds to the harmonic mean of the two values but other values weight recall beta times more important than precision.

estimator
One of: "binary", "macro", "macro_weighted", or "micro" to specify the type of averaging to be done. "binary" is only relevant for the two class case. The other three are general methods for calculating multiclass metrics. The default will automatically choose "binary" or "macro" based on estimate.

na_rm
A logical value indicating whether NA values should be stripped before the computation proceeds.

Value

A tibble with columns .metric, .estimator, and .estimate and 1 row of values.

For grouped data frames, the number of rows returned will be the same as the number of groups.

For f_meas_vec(), a single numeric value (or NA).

Details

The measure "F" is a combination of precision and recall (see below).

Relevant level

There is no common convention on which factor level should automatically be considered the "event" or "positive" result. In yardstick, the default is to use the first level. To change this, a global option called yardstick.event_first is set to TRUE when the package is loaded. This can be changed to FALSE if the last level of the factor is considered the level of interest. For multiclass extensions involving one-vs-all comparisons (such as macro averaging), this option is ignored and the "one" level is always the relevant result.

Multiclass

Macro, micro, and macro-weighted averaging is available for this metric. The default is to select macro averaging if a truth factor with more than 2 levels is provided. Otherwise, a standard binary calculation is done. See vignette("multiclass", "yardstick") for more information.

Implementation

Suppose a 2x2 table with notation:

Reference

Predicted	Relevant

Irrelevant	Relevant	A

B	Irrelevant	C

D

The formulas used here are:

$$recall = A/(A+C)$$ $$precision = A/(A+B)$$ $$F_meas_\beta = (1+\beta^2) * precision * recall/((\beta^2 * precision)+recall)$$

See the references for discussions of the statistics.

References

Buckland, M., & Gey, F. (1994). The relationship between Recall and Precision. Journal of the American Society for Information Science, 45(1), 12-19.

Powers, D. (2007). Evaluation: From Precision, Recall and F Factor to ROC, Informedness, Markedness and Correlation. Technical Report SIE-07-001, Flinders University

See also

Examples

# Two class
data("two_class_example")
f_meas(two_class_example, truth, predicted)

#> # A tibble: 1 x 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 f_meas  binary         0.849

# Multiclass
library(dplyr)
data(hpc_cv)

hpc_cv %>%
  filter(Resample == "Fold01") %>%
  f_meas(obs, pred)

#> # A tibble: 1 x 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 f_meas  macro          0.563

# Groups are respected
hpc_cv %>%
  group_by(Resample) %>%
  f_meas(obs, pred)

#> # A tibble: 10 x 4
#>    Resample .metric .estimator .estimate
#>    <chr>    <chr>   <chr>          <dbl>
#>  1 Fold01   f_meas  macro          0.563
#>  2 Fold02   f_meas  macro          0.542
#>  3 Fold03   f_meas  macro          0.641
#>  4 Fold04   f_meas  macro          0.593
#>  5 Fold05   f_meas  macro          0.570
#>  6 Fold06   f_meas  macro          0.554
#>  7 Fold07   f_meas  macro          0.516
#>  8 Fold08   f_meas  macro          0.601
#>  9 Fold09   f_meas  macro          0.555
#> 10 Fold10   f_meas  macro          0.560

# Weighted macro averaging
hpc_cv %>%
  group_by(Resample) %>%
  f_meas(obs, pred, estimator = "macro_weighted")

#> # A tibble: 10 x 4
#>    Resample .metric .estimator     .estimate
#>    <chr>    <chr>   <chr>              <dbl>
#>  1 Fold01   f_meas  macro_weighted     0.696
#>  2 Fold02   f_meas  macro_weighted     0.684
#>  3 Fold03   f_meas  macro_weighted     0.739
#>  4 Fold04   f_meas  macro_weighted     0.689
#>  5 Fold05   f_meas  macro_weighted     0.692
#>  6 Fold06   f_meas  macro_weighted     0.673
#>  7 Fold07   f_meas  macro_weighted     0.646
#>  8 Fold08   f_meas  macro_weighted     0.701
#>  9 Fold09   f_meas  macro_weighted     0.652
#> 10 Fold10   f_meas  macro_weighted     0.680

# Vector version
f_meas_vec(two_class_example$truth, two_class_example$predicted)

#> [1] 0.8485981

# Making Class2 the "relevant" level
options(yardstick.event_first = FALSE)
f_meas_vec(two_class_example$truth, two_class_example$predicted)

#> [1] 0.8258065

options(yardstick.event_first = TRUE)