Index of ideality of correlation

Calculate the index of ideality of correlation. This metric has been studied in QSPR/QSAR models as a good criterion for the predictive potential of these models. It is highly dependent on the correlation coefficient as well as the mean absolute error.

Note the application of IIC is useless under two conditions:

When the negative mean absolute error and positive mean absolute error are both zero.

When the outliers are symmetric. Since outliers are context dependent, please use your own checks to validate whether this restriction holds and whether the resulting IIC has interpretative value.

The IIC is seen as an alternative to the traditional correlation coefficient and is in the same units as the original data.

iic(data, ...)

# S3 method for data.frame
iic(data, truth, estimate, na_rm = TRUE, ...)

iic_vec(truth, estimate, na_rm = TRUE, ...)

Arguments

data
A data.frame containing the truth and estimate columns.

...
Not currently used.

truth
The column identifier for the true results (that is numeric). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a numeric vector.

estimate
The column identifier for the predicted results (that is also numeric). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a numeric vector.

na_rm
A logical value indicating whether NA values should be stripped before the computation proceeds.

Value

A tibble with columns .metric, .estimator, and .estimate and 1 row of values.

For grouped data frames, the number of rows returned will be the same as the number of groups.

For iic_vec(), a single numeric value (or NA).

References

Toropova, A. and Toropov, A. (2017). "The index of ideality of correlation. A criterion of predictability of QSAR models for skin permeability?" Science of the Total Environment. 586: 466-472.

See also

Examples

# Supply truth and predictions as bare column names
iic(solubility_test, solubility, prediction)

#> # A tibble: 1 x 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 iic     standard       0.890

library(dplyr)

set.seed(1234)
size <- 100
times <- 10

# create 10 resamples
solubility_resampled <- bind_rows(
  replicate(
    n = times,
    expr = sample_n(solubility_test, size, replace = TRUE),
    simplify = FALSE
  ),
  .id = "resample"
)

# Compute the metric by group
metric_results <- solubility_resampled %>%
  group_by(resample) %>%
  iic(solubility, prediction)

metric_results

#> # A tibble: 10 x 4
#>    resample .metric .estimator .estimate
#>    <chr>    <chr>   <chr>          <dbl>
#>  1 1        iic     standard       0.730
#>  2 10       iic     standard       0.731
#>  3 2        iic     standard       0.906
#>  4 3        iic     standard       0.877
#>  5 4        iic     standard       0.732
#>  6 5        iic     standard       0.821
#>  7 6        iic     standard       0.896
#>  8 7        iic     standard       0.867
#>  9 8        iic     standard       0.881
#> 10 9        iic     standard       0.748

# Resampled mean estimate
metric_results %>%
  summarise(avg_estimate = mean(.estimate))

#> # A tibble: 1 x 1
#>   avg_estimate
#>          <dbl>
#> 1        0.819