Survival Analysis Results

Survival Analysis Results

Value

lung_surv

a data frame

Details

These data contain plausible results from applying predictive survival models to the lung data set using the censored package.

Examples

data(lung_surv)
str(lung_surv)
#> tibble [150 × 14] (S3: tbl_df/tbl/data.frame)
#>  $ surv_obj      : 'Surv' num [1:150, 1:2] 455  455  455  210  210  210  1022 1022 1022 310  ...
#>   ..- attr(*, "type")= chr "right"
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "time" "status"
#>  $ .time         : num [1:150] 100 500 1000 100 500 1000 100 500 1000 100 ...
#>  $ .pred_survival: num [1:150] 0.9274 0.3666 0.0469 0.8596 0.1334 ...
#>  $ ipcw          : num [1:150] 1 1 1 1 1 1 1 1 3 1 ...
#>  $ inst          : num [1:150] 3 3 3 5 5 5 12 12 12 7 ...
#>  $ time          : num [1:150] 455 455 455 210 210 ...
#>  $ status        : num [1:150] 2 2 2 2 2 2 1 1 1 2 ...
#>  $ age           : num [1:150] 68 68 68 57 57 57 74 74 74 68 ...
#>  $ sex           : num [1:150] 1 1 1 1 1 1 1 1 1 2 ...
#>  $ ph.ecog       : num [1:150] 0 0 0 1 1 1 1 1 1 2 ...
#>  $ ph.karno      : num [1:150] 90 90 90 90 90 90 50 50 50 70 ...
#>  $ pat.karno     : num [1:150] 90 90 90 60 60 60 80 80 80 60 ...
#>  $ meal.cal      : num [1:150] 1225 1225 1225 1150 1150 ...
#>  $ wt.loss       : num [1:150] 15 15 15 11 11 11 0 0 0 10 ...

# `surv_obj` is a `Surv()` object