Mean absolute error

Calculate the mean absolute error. This metric is in the same units as the original data.

mae(data, ...)

# S3 method for data.frame
mae(data, truth, estimate, na_rm = TRUE, ...)

mae_vec(truth, estimate, na_rm = TRUE, ...)

Arguments

data
A data.frame containing the truth and estimate columns.

...
Not currently used.

truth
The column identifier for the true results (that is numeric). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a numeric vector.

estimate
The column identifier for the predicted results (that is also numeric). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a numeric vector.

na_rm
A logical value indicating whether NA values should be stripped before the computation proceeds.

Value

A tibble with columns .metric, .estimator, and .estimate and 1 row of values.

For grouped data frames, the number of rows returned will be the same as the number of groups.

For mae_vec(), a single numeric value (or NA).

See also

Author

Max Kuhn

Examples

# Supply truth and predictions as bare column names
mae(solubility_test, solubility, prediction)
#> # A tibble: 1 × 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 mae     standard       0.545

library(dplyr)

set.seed(1234)
size <- 100
times <- 10

# create 10 resamples
solubility_resampled <- bind_rows(
  replicate(
    n = times,
    expr = sample_n(solubility_test, size, replace = TRUE),
    simplify = FALSE
  ),
  .id = "resample"
)

# Compute the metric by group
metric_results <- solubility_resampled %>%
  group_by(resample) %>%
  mae(solubility, prediction)

metric_results
#> # A tibble: 10 × 4
#>    resample .metric .estimator .estimate
#>    <chr>    <chr>   <chr>          <dbl>
#>  1 1        mae     standard       0.512
#>  2 10       mae     standard       0.504
#>  3 2        mae     standard       0.545
#>  4 3        mae     standard       0.496
#>  5 4        mae     standard       0.587
#>  6 5        mae     standard       0.500
#>  7 6        mae     standard       0.627
#>  8 7        mae     standard       0.566
#>  9 8        mae     standard       0.473
#> 10 9        mae     standard       0.540

# Resampled mean estimate
metric_results %>%
  summarise(avg_estimate = mean(.estimate))
#> # A tibble: 1 × 1
#>   avg_estimate
#>          <dbl>
#> 1        0.535