Combine metric functions

metric_set() allows you to combine multiple metric functions together into a new function that calculates all of them at once.

metric_set(...)

Arguments

...
The bare names of the functions to be included in the metric set.

Details

There are currently no checks in place to ensure that all of the metric functions calculate the same kind of metric (regression vs classification).

See also

Examples

library(dplyr)

# Multiple regression metrics
multi_metric <- metric_set(rmse, rsq, ccc)

multi_metric(solubility_test, truth = solubility, estimate = prediction)

#> # A tibble: 3 x 2
#>   .metric .estimate
#>   <chr>       <dbl>
#> 1 rmse        0.722
#> 2 rsq         0.879
#> 3 ccc         0.934

# Groups are respected on the new metric function
class_metrics <- metric_set(accuracy, kap)

hpc_cv %>%
  group_by(Resample) %>%
  class_metrics(obs, pred)

#> # A tibble: 20 x 3
#>    Resample .metric  .estimate
#>    <chr>    <chr>        <dbl>
#>  1 Fold01   accuracy     0.726
#>  2 Fold02   accuracy     0.712
#>  3 Fold03   accuracy     0.758
#>  4 Fold04   accuracy     0.712
#>  5 Fold05   accuracy     0.712
#>  6 Fold06   accuracy     0.697
#>  7 Fold07   accuracy     0.675
#>  8 Fold08   accuracy     0.721
#>  9 Fold09   accuracy     0.673
#> 10 Fold10   accuracy     0.699
#> 11 Fold01   kap          0.533
#> 12 Fold02   kap          0.512
#> 13 Fold03   kap          0.594
#> 14 Fold04   kap          0.511
#> 15 Fold05   kap          0.514
#> 16 Fold06   kap          0.486
#> 17 Fold07   kap          0.454
#> 18 Fold08   kap          0.531
#> 19 Fold09   kap          0.454
#> 20 Fold10   kap          0.492

# If you need to set options for certain metrics,
# do so ahead of time with purrr::partial()
library(purrr)

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:yardstick’:
#> 
#>     flatten

ccc_with_bias <- partial(ccc, bias = TRUE)

multi_metric2 <- metric_set(rmse, rsq, ccc_with_bias)

multi_metric2(solubility_test, truth = solubility, estimate = prediction)

#> # A tibble: 3 x 2
#>   .metric .estimate
#>   <chr>       <dbl>
#> 1 rmse        0.722
#> 2 rsq         0.879
#> 3 ccc         0.937