Negative predictive value

These functions calculate the npv() (negative predictive value) of a measurement system compared to a reference result (the "truth" or gold standard). Highly related functions are spec(), sens(), and ppv().

npv(data, ...)

# S3 method for data.frame
npv(
  data,
  truth,
  estimate,
  prevalence = NULL,
  estimator = NULL,
  na_rm = TRUE,
  ...
)

npv_vec(
  truth,
  estimate,
  prevalence = NULL,
  estimator = NULL,
  na_rm = TRUE,
  ...
)

Arguments

data
Either a data.frame containing the truth and estimate columns, or a table/matrix where the true class results should be in the columns of the table.

...
Not currently used.

truth
The column identifier for the true class results (that is a factor). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a factor vector.

estimate
The column identifier for the predicted class results (that is also factor). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a factor vector.

prevalence
A numeric value for the rate of the "positive" class of the data.

estimator
One of: "binary", "macro", "macro_weighted", or "micro" to specify the type of averaging to be done. "binary" is only relevant for the two class case. The other three are general methods for calculating multiclass metrics. The default will automatically choose "binary" or "macro" based on estimate.

na_rm
A logical value indicating whether NA values should be stripped before the computation proceeds.

Value

A tibble with columns .metric, .estimator, and .estimate and 1 row of values.

For grouped data frames, the number of rows returned will be the same as the number of groups.

For npv_vec(), a single numeric value (or NA).

Details

The positive predictive value (ppv()) is defined as the percent of predicted positives that are actually positive while the negative predictive value (npv()) is defined as the percent of negative positives that are actually negative.

Relevant Level

There is no common convention on which factor level should automatically be considered the "event" or "positive" result. In yardstick, the default is to use the first level. To change this, a global option called yardstick.event_first is set to TRUE when the package is loaded. This can be changed to FALSE if the last level of the factor is considered the level of interest by running: options(yardstick.event_first = FALSE). For multiclass extensions involving one-vs-all comparisons (such as macro averaging), this option is ignored and the "one" level is always the relevant result.

Multiclass

Macro, micro, and macro-weighted averaging is available for this metric. The default is to select macro averaging if a truth factor with more than 2 levels is provided. Otherwise, a standard binary calculation is done. See vignette("multiclass", "yardstick") for more information.

Implementation

Suppose a 2x2 table with notation:

Reference

Predicted	Positive	Negative

Positive	A	B

Negative	C	D

The formulas used here are:

$$Sensitivity = A/(A+C)$$ $$Specificity = D/(B+D)$$ $$Prevalence = (A+C)/(A+B+C+D)$$ $$PPV = (Sensitivity * Prevalence) / ((Sensitivity * Prevalence) + ((1-Specificity) * (1-Prevalence)))$$ $$NPV = (Specificity * (1-Prevalence)) / (((1-Sensitivity) * Prevalence) + ((Specificity) * (1-Prevalence)))$$

See the references for discussions of the statistics.

References

Altman, D.G., Bland, J.M. (1994) ``Diagnostic tests 2: predictive values,'' British Medical Journal, vol 309, 102.

See also

Examples

# Two class
data("two_class_example")
npv(two_class_example, truth, predicted)

#> # A tibble: 1 x 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 npv     binary         0.861

# Multiclass
library(dplyr)
data(hpc_cv)

hpc_cv %>%
  filter(Resample == "Fold01") %>%
  npv(obs, pred)

#> # A tibble: 1 x 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 npv     macro          0.906

# Groups are respected
hpc_cv %>%
  group_by(Resample) %>%
  npv(obs, pred)

#> # A tibble: 10 x 4
#>    Resample .metric .estimator .estimate
#>    <chr>    <chr>   <chr>          <dbl>
#>  1 Fold01   npv     macro          0.906
#>  2 Fold02   npv     macro          0.901
#>  3 Fold03   npv     macro          0.917
#>  4 Fold04   npv     macro          0.897
#>  5 Fold05   npv     macro          0.897
#>  6 Fold06   npv     macro          0.892
#>  7 Fold07   npv     macro          0.882
#>  8 Fold08   npv     macro          0.902
#>  9 Fold09   npv     macro          0.879
#> 10 Fold10   npv     macro          0.890

# Weighted macro averaging
hpc_cv %>%
  group_by(Resample) %>%
  npv(obs, pred, estimator = "macro_weighted")

#> # A tibble: 10 x 4
#>    Resample .metric .estimator     .estimate
#>    <chr>    <chr>   <chr>              <dbl>
#>  1 Fold01   npv     macro_weighted     0.896
#>  2 Fold02   npv     macro_weighted     0.890
#>  3 Fold03   npv     macro_weighted     0.905
#>  4 Fold04   npv     macro_weighted     0.878
#>  5 Fold05   npv     macro_weighted     0.878
#>  6 Fold06   npv     macro_weighted     0.871
#>  7 Fold07   npv     macro_weighted     0.853
#>  8 Fold08   npv     macro_weighted     0.885
#>  9 Fold09   npv     macro_weighted     0.845
#> 10 Fold10   npv     macro_weighted     0.864

# Vector version
npv_vec(two_class_example$truth, two_class_example$predicted)

#> [1] 0.8609865

# Making Class2 the "relevant" level
options(yardstick.event_first = FALSE)
npv_vec(two_class_example$truth, two_class_example$predicted)

#> [1] 0.8194946

options(yardstick.event_first = TRUE)