Numeric metrics

Numeric metrics evaluate regression predictions where both truth and estimate are numeric. These metrics measure how close predicted values are to the true values.

Input requirements

truth: numeric

estimate: numeric

Available metrics

ccc()

Direction: maximize. Range: [-1, 1]

huber_loss()

Direction: minimize. Range: [0, Inf]

huber_loss_pseudo()

Direction: minimize. Range: [0, Inf]

iic()

Direction: maximize. Range: [-1, 1]

mae()

Direction: minimize. Range: [0, Inf]

mape()

Direction: minimize. Range: [0, Inf]

mase()

Direction: minimize. Range: [0, Inf]

mpe()

Direction: zero. Range: [-Inf, Inf]

msd()

Direction: zero. Range: [-Inf, Inf]

poisson_log_loss()

Direction: minimize. Range: [0, Inf]

rmse()

Direction: minimize. Range: [0, Inf]

rmse_relative()

Direction: minimize. Range: [0, Inf]

rpd()

Direction: maximize. Range: [0, Inf]

rpiq()

Direction: maximize. Range: [0, Inf]

rsq()

Direction: maximize. Range: [-Inf, 1]

rsq_trad()

Direction: maximize. Range: [0, 1]

smape()

Direction: minimize. Range: [0, 100]

See also

Examples

data("solubility_test")

head(solubility_test)
#>   solubility prediction
#> 1       0.93  0.3677522
#> 2       0.85 -0.1503220
#> 3       0.81 -0.5051844
#> 4       0.74  0.5398116
#> 5       0.61 -0.4792718
#> 6       0.58  0.7377222

rmse(solubility_test, solubility, prediction)
#> # A tibble: 1 × 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 rmse    standard       0.722