Metrics Based on Class Probabilities

These functions compute the areas under the receiver operating characteristic (ROC) curve (roc_auc), the precision-recall curve (pr_auc), or the multinomial log loss (mnLogLoss).

roc_auc(data, ...)

# S3 method for data.frame
roc_auc(data, truth, ..., options = list(),
  na.rm = TRUE)

pr_auc(data, ...)

# S3 method for data.frame
pr_auc(data, truth, ..., na.rm = TRUE)

mnLogLoss(data, ...)

# S3 method for data.frame
mnLogLoss(data, truth, ..., na.rm = TRUE, sum = FALSE)

Arguments

data
A data frame with the relevant columns.

...
A set of unquoted column names or one or more dplyr selector functions to choose which variables contain the class probabilities. See the examples below. For roc_auc and pr_auc, only one value is required. If more are given, the functions will try to match the column name to the appropriate factor level of truth. If this doesn't work, an error is thrown. For mnLogLoss, there should be as many columns as factor levels of truth. It is assumed that they are in the same order as the factor levels.

truth
The column identifier for the true class results (that is a factor). This should an unquoted column name although this argument is passed by expression and support quasiquotation (you can unquote column names or column positions).

options
Options to pass to roc() such as direction or smooth. These options should not include response, predictor, or levels.

na.rm
A logical value indicating whether NA values should be stripped before the computation proceeds

sum
A logical. Should the sum of the likelihood contrinbutions be returned (instead of the mean value)?

Value

A number between 0 and 1 (or NA) for roc_auc or pr_auc. For mnLogLoss a number or NA.

Details

There is no common convention on which factor level should automatically be considered the "relevant" or "positive" results. In yardstick, the default is to use the first level. To change this, a global option called yardstick.event_first is set to TRUE when the package is loaded. This can be changed to FALSE if the last level of the factor is considered the level of interest.

See also

conf_mat(), summary.conf_mat(), recall(), mcc()

Examples

library(tidyselect)

#> 
#> Attaching package: ‘tidyselect’

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     contains, ends_with, everything, matches, num_range, one_of,
#>     starts_with

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

data("two_class_example")
prob_cols <- levels(two_class_example$truth)

roc_auc(two_class_example, truth = truth, Class1)

#> [1] 0.9393139

# warning is issued here because 2 columns are selected:
roc_auc(two_class_example, truth, starts_with("Class"))

#> Warning: Multiple columns were specified for class probabilities; Class1 will be used.

#> [1] 0.9393139

# passing options via a list and _not_ `...`
roc_auc(two_class_example, truth = "truth", Class1,
        options = list(smooth = TRUE))

#> [1] 0.9421511

pr_auc(two_class_example, truth, prob_cols)

#> Warning: Multiple columns were specified for class probabilities; Class1 will be used.

#> [1] 0.9425707

mnLogLoss(two_class_example, truth, starts_with("Class"))

#> [1] -0.3283096

# or
mnLogLoss(two_class_example, truth, !! prob_cols)

#> [1] -0.3283096