Classification Metrics on Predited Classes

Accuracy is the proportion of the data that are predicted correctly. Kappa is a similar measure but is normalized by the accuracy that would be expected by chance alone and is very useful when one or more classes have large frequency distributions.

accuracy(data, ...)

# S3 method for data.frame
accuracy(data, truth, estimate, na.rm = TRUE, ...)

# S3 method for table
accuracy(data, ...)

# S3 method for matrix
accuracy(data, ...)

kap(data, ...)

# S3 method for data.frame
kap(data, truth, estimate, na.rm = TRUE, ...)

# S3 method for table
kap(data, ...)

# S3 method for matrix
kap(data, ...)

Arguments

data
For the default functions, a factor containing the discrete measurements. For the table or matrix functions, a table or matrix object, respectively, where the true class results should be in the columns of the table.

...
Not currently used.

truth
The column identifier for the true class results (that is a factor). This should an unquoted column name although this argument is passed by expression and support quasiquotation (you can unquote column names or column positions).

estimate
The column identifier for the predicted class results (that is also factor). As with truth this can be specified different ways but the primary method is to use an unquoted variable name.

na.rm
A logical value indicating whether NA values should be stripped before the computation proceeds

Value

A numeric value.

References

Cohen, J. (1960). "A coefficient of agreement for nominal scales". Educational and Psychological Measurement. 20 (1): 37–46.

See also

Examples

library(dplyr)
data("hpc_cv")

# The observed and predicted classes from a single
# assessment set (i.e. fold)
fold_1 <- hpc_cv %>%
  filter(Resample == "Fold01")

fold_1 %>% conf_mat(truth = obs, estimate = pred)

#>           Truth
#> Prediction  VF   F   M   L
#>         VF 166  33   8   1
#>         F   11  71  24   7
#>         M    0   3   5   3
#>         L    0   1   4  10

fold_1 %>% accuracy(truth = obs, estimate = pred)

#> [1] 0.726

fold_1 %>% kap(truth = obs, estimate = pred)

#> [1] 0.533

fold_1 %>% metrics(truth = obs, estimate = pred)

#> # A tibble: 1 x 2
#>   accuracy kappa
#>      <dbl> <dbl>
#> 1    0.726 0.533