Root mean squared error

Calculate the root mean squared error. rmse() is a metric that is in the same units as the original data.

rmse(data, ...)

# S3 method for data.frame
rmse(data, truth, estimate, na_rm = TRUE, ...)

rmse_vec(truth, estimate, na_rm = TRUE, ...)

Arguments

data
A data.frame containing the truth and estimate columns.

...
Not currently used.

truth
The column identifier for the true results (that is numeric). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a numeric vector.

estimate
The column identifier for the predicted results (that is also numeric). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a numeric vector.

na_rm
A logical value indicating whether NA values should be stripped before the computation proceeds.

Value

A tibble with columns .metric, .estimator, and .estimate and 1 row of values.

For grouped data frames, the number of rows returned will be the same as the number of groups.

For rmse_vec(), a single numeric value (or NA).

See also

Examples

# Supply truth and predictions as bare column names
rmse(solubility_test, solubility, prediction)

#> # A tibble: 1 x 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 rmse    standard       0.722

library(dplyr)

set.seed(1234)
size <- 100
times <- 10

# create 10 resamples
solubility_resampled <- bind_rows(
  replicate(
    n = times,
    expr = sample_n(solubility_test, size, replace = TRUE),
    simplify = FALSE
  ),
  .id = "resample"
)

# Compute the metric by group
metric_results <- solubility_resampled %>%
  group_by(resample) %>%
  rmse(solubility, prediction)

metric_results

#> # A tibble: 10 x 4
#>    resample .metric .estimator .estimate
#>    <chr>    <chr>   <chr>          <dbl>
#>  1 1        rmse    standard       0.715
#>  2 10       rmse    standard       0.673
#>  3 2        rmse    standard       0.716
#>  4 3        rmse    standard       0.644
#>  5 4        rmse    standard       0.737
#>  6 5        rmse    standard       0.675
#>  7 6        rmse    standard       0.807
#>  8 7        rmse    standard       0.801
#>  9 8        rmse    standard       0.635
#> 10 9        rmse    standard       0.692

# Resampled mean estimate
metric_results %>%
  summarise(avg_estimate = mean(.estimate))

#> # A tibble: 1 x 1
#>   avg_estimate
#>          <dbl>
#> 1        0.710