Sensitivity

These functions calculate the sens() (sensitivity) of a measurement system compared to a reference result (the "truth" or gold standard). Highly related functions are spec(), ppv(), and npv().

Usage

sens(data, ...)

# S3 method for class 'data.frame'
sens(
  data,
  truth,
  estimate,
  estimator = NULL,
  na_rm = TRUE,
  case_weights = NULL,
  event_level = yardstick_event_level(),
  ...
)

sens_vec(
  truth,
  estimate,
  estimator = NULL,
  na_rm = TRUE,
  case_weights = NULL,
  event_level = yardstick_event_level(),
  ...
)

sensitivity(data, ...)

# S3 method for class 'data.frame'
sensitivity(
  data,
  truth,
  estimate,
  estimator = NULL,
  na_rm = TRUE,
  case_weights = NULL,
  event_level = yardstick_event_level(),
  ...
)

sensitivity_vec(
  truth,
  estimate,
  estimator = NULL,
  na_rm = TRUE,
  case_weights = NULL,
  event_level = yardstick_event_level(),
  ...
)

Arguments

data

Either a data.frame containing the columns specified by the truth and estimate arguments, or a table/matrix where the true class results should be in the columns of the table.

...

Not currently used.

truth

The column identifier for the true class results (that is a factor). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a factor vector.

estimate

The column identifier for the predicted class results (that is also factor). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a factor vector.

estimator

One of: "binary", "macro", "macro_weighted", or "micro" to specify the type of averaging to be done. "binary" is only relevant for the two class case. The other three are general methods for calculating multiclass metrics. The default will automatically choose "binary" or "macro" based on estimate.

na_rm

A logical value indicating whether NA values should be stripped before the computation proceeds.

case_weights

The optional column identifier for case weights. This should be an unquoted column name that evaluates to a numeric column in data. For _vec() functions, a numeric vector, hardhat::importance_weights(), or hardhat::frequency_weights().

event_level

A single string. Either "first" or "second" to specify which level of truth to consider as the "event". This argument is only applicable when estimator = "binary". The default uses an internal helper that defaults to "first".

Value

A tibble with columns .metric, .estimator, and .estimate and 1 row of values.

For grouped data frames, the number of rows returned will be the same as the number of groups.

For sens_vec(), a single numeric value (or NA).

Details

The sensitivity (sens()) is defined as the proportion of positive results out of the number of samples which were actually positive. For positive observations, the proportion of model predictions that correctly predicted positive.

When the denominator of the calculation is 0, sensitivity is undefined. This happens when both # true_positive = 0 and # false_negative = 0 are true, which mean that there were no true events. When computing binary sensitivity, a NA value will be returned with a warning. When computing multiclass sensitivity, the individual NA values will be removed, and the computation will procede, with a warning.

Suppose a 2x2 table with notation:

Reference

Predicted	Positive	Negative

Positive	A	B

Negative	C	D

The formula used here is:

$$\text{Sensitivity} = \frac{A}{A + C}$$

Sensitivity is a metric that should be maximized. The output ranges from 0 to 1, with 1 indicating that all actual positives were predicted as positive.

Relevant Level

There is no common convention on which factor level should automatically be considered the "event" or "positive" result when computing binary classification metrics. In yardstick, the default is to use the first level. To alter this, change the argument event_level to "second" to consider the last level of the factor the level of interest. For multiclass extensions involving one-vs-all comparisons (such as macro averaging), this option is ignored and the "one" level is always the relevant result.

Multiclass

Macro, micro, and macro-weighted averaging is available for this metric. The default is to select macro averaging if a truth factor with more than 2 levels is provided. Otherwise, a standard binary calculation is done. See vignette("multiclass", "yardstick") for more information.

References

Altman, D.G., Bland, J.M. (1994) “Diagnostic tests 1: sensitivity and specificity,” British Medical Journal, vol 308, 1552.

See also

Author

Max Kuhn

Examples

# Two class
data("two_class_example")
sens(two_class_example, truth, predicted)
#> # A tibble: 1 × 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 sens    binary         0.880

# Multiclass
library(dplyr)
data(hpc_cv)

hpc_cv |>
  filter(Resample == "Fold01") |>
  sens(obs, pred)
#> # A tibble: 1 × 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 sens    macro          0.548

# Groups are respected
hpc_cv |>
  group_by(Resample) |>
  sens(obs, pred)
#> # A tibble: 10 × 4
#>    Resample .metric .estimator .estimate
#>    <chr>    <chr>   <chr>          <dbl>
#>  1 Fold01   sens    macro          0.548
#>  2 Fold02   sens    macro          0.541
#>  3 Fold03   sens    macro          0.634
#>  4 Fold04   sens    macro          0.570
#>  5 Fold05   sens    macro          0.550
#>  6 Fold06   sens    macro          0.540
#>  7 Fold07   sens    macro          0.531
#>  8 Fold08   sens    macro          0.584
#>  9 Fold09   sens    macro          0.568
#> 10 Fold10   sens    macro          0.537

# Weighted macro averaging
hpc_cv |>
  group_by(Resample) |>
  sens(obs, pred, estimator = "macro_weighted")
#> # A tibble: 10 × 4
#>    Resample .metric .estimator     .estimate
#>    <chr>    <chr>   <chr>              <dbl>
#>  1 Fold01   sens    macro_weighted     0.726
#>  2 Fold02   sens    macro_weighted     0.712
#>  3 Fold03   sens    macro_weighted     0.758
#>  4 Fold04   sens    macro_weighted     0.712
#>  5 Fold05   sens    macro_weighted     0.712
#>  6 Fold06   sens    macro_weighted     0.697
#>  7 Fold07   sens    macro_weighted     0.675
#>  8 Fold08   sens    macro_weighted     0.721
#>  9 Fold09   sens    macro_weighted     0.673
#> 10 Fold10   sens    macro_weighted     0.699

# Vector version
sens_vec(
  two_class_example$truth,
  two_class_example$predicted
)
#> [1] 0.879845

# Making Class2 the "relevant" level
sens_vec(
  two_class_example$truth,
  two_class_example$predicted,
  event_level = "second"
)
#> [1] 0.7933884