Summary Statistics for Confusion Matrices

Various statistical summaries of confusion matrices are produced and returned in a easily used format. These potentially include those shown in the help pages for sens(), recall(), and accuracy().

# S3 method for conf_mat
summary(object, prevalence = NULL, beta = 1, ...)

Arguments

object
An object of class conf_mat().

prevalence
A number in (0, 1) for the prevalence (i.e. prior) of the event. If left to the default, the data are used to derive this value.

beta
A numeric value used to weight precision and recall for f_meas().

...
Not currently used.

Value

A tibble. Note that if the argument prevalence was used, the value reported in the tibble reflects the argument value and not the observed rate of events.

Details

There is no common convention on which factor level should automatically be considered the "event" or "positive" results. In yardstick, the default is to use the first level. To change this, a global option called yardstick.event_first is set to TRUE when the package is loaded. This can be changed to FALSE if the last level of the factor is considered the level of interest.

Examples

data("two_class_example")

cmat <- conf_mat(two_class_example, truth = "truth", estimate = "predicted")
summary(cmat)

#> # A tibble: 14 x 2
#>    .metric              .estimate
#>    <chr>                    <dbl>
#>  1 accuracy                 0.838
#>  2 kap                      0.675
#>  3 sens                     0.880
#>  4 spec                     0.793
#>  5 prevalence               0.516
#>  6 ppv                      0.819
#>  7 npv                      0.861
#>  8 mcc                      0.677
#>  9 j_index                  0.673
#> 10 bal_accuracy             0.837
#> 11 detection_prevalence     0.554
#> 12 precision                0.819
#> 13 recall                   0.880
#> 14 f_meas                   0.849

summary(cmat, prevalence = 0.70)

#> # A tibble: 14 x 2
#>    .metric              .estimate
#>    <chr>                    <dbl>
#>  1 accuracy                 0.838
#>  2 kap                      0.675
#>  3 sens                     0.880
#>  4 spec                     0.793
#>  5 prevalence               0.7  
#>  6 ppv                      0.909
#>  7 npv                      0.739
#>  8 mcc                      0.677
#>  9 j_index                  0.673
#> 10 bal_accuracy             0.837
#> 11 detection_prevalence     0.554
#> 12 precision                0.819
#> 13 recall                   0.880
#> 14 f_meas                   0.849

library(dplyr)
library(purrr)
library(tidyr)
data("hpc_cv")

# Compute statistics per resample then summarize
all_metrics <- hpc_cv %>%
  group_by(Resample) %>%
  conf_mat(obs, pred) %>%
  mutate(summary_tbl = map(conf_mat, summary)) %>%
  unnest(summary_tbl)

all_metrics %>%
  group_by(Resample) %>%
  summarise(
    mean = mean(.estimate, na.rm = TRUE),
    sd = sd(.estimate, na.rm = TRUE)
  )

#> # A tibble: 10 x 3
#>    Resample  mean    sd
#>    <chr>    <dbl> <dbl>
#>  1 Fold01   0.630 0.136
#>  2 Fold02   0.612 0.141
#>  3 Fold03   0.676 0.116
#>  4 Fold04   0.611 0.142
#>  5 Fold05   0.613 0.140
#>  6 Fold06   0.592 0.149
#>  7 Fold07   0.565 0.156
#>  8 Fold08   0.626 0.135
#>  9 Fold09   0.564 0.155
#> 10 Fold10   0.596 0.146