yardstick: Tidy Characterizations of Model Performance

Tidy tools for quantifying how well model fits to a data set such as confusion matrices, class probability curve summaries, and regression metrics (e.g., RMSE).

See also

Author

Maintainer: Emil Hvitfeldt emil.hvitfeldt@posit.co (ORCID)

Authors:

Max Kuhn max@posit.co

Davis Vaughan davis@posit.co

Other contributors:

Posit Software, PBC (ROR) [copyright holder, funder]